refactor(SinglePaymentService): 공통 결제 로직 분리
